Refactored cloudprober.go and client.go
