+++ cloudprober/cloudprober.go
@@ -1,47 +1,37 @@
-package main
+// Copyright 2020 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// Author: Evan Spendlove (@evanSpendlove)
+//
+// cloudprober.go initialises and runs a cloudprober instance using a custom configuration file.
+
+package cloudprober
 
 import (
 	"context"
-	"flag"
-	"fmt"
-	"io/ioutil"
 
 	"github.com/golang/glog"
 	"github.com/google/cloudprober"
-	"github.com/googleinterns/step224-2020/cloudprober/myprobe"
-	"github.com/google/cloudprober/probes"
 	"github.com/google/cloudprober/web"
 )
 
-var (
-	configFile = flag.String("config_file", "", "Config file")
-)
-
-func configFileToString(fileName string) string {
-	b, err := ioutil.ReadFile(fileName)
-	if err != nil {
-		fmt.Printf("Failed to read the config file: %v", err)
-	}
-	return string(b)
-}
-
-func getConfig() string {
-	if *configFile != "" {
-		return configFileToString(*configFile)
-	} else {
-		fmt.Println("Config file not found")
-		return ""
-	}
-}
-
-func main() {
-	flag.Parse()
-
-	// Register stubby probe type
-	probes.RegisterProbeType(int(myprobe.E_RedisProbe.Field), // TODO: CHANGE THIS
-		func() probes.Probe { return &myprobe.Probe{} })
-
-	err := cloudprober.InitFromConfig(getConfig())
+// InitialiseCloudproberFromConfig initialises Cloudprober from the config passed as an argument.
+// It then sets up the web UI for Cloudprober, and starts running Cloudprober.
+// The config passed as a string should be the contents of a Cloudprober config file, not a file path or filename.
+// For most instances, this is likely just "grpc_port=9314" as this is the only required value in the config.
+func InitialiseCloudproberFromConfig(config string) {
+	err := cloudprober.InitFromConfig(config)
 	if err != nil {
 		glog.Exitf("Error initializing cloudprober. Err: %v", err)
 	}
@@ -49,8 +39,5 @@ func main() {
 	// web.Init sets up web UI for cloudprober.
 	web.Init()
 
-	cloudprober.Start(context.Background())
-
-	// Wait forever
-	select {}
+	cloudprober.Start(context.Background()) // Start running Cloudprober instance
 }
