+++ cloudprober/myprober.cfg
@@ -1,16 +1,1 @@
-probe {
-  name: "redis_set"
-  type: EXTENSION
-  interval_msec: 10000
-  timeout_msec: 5000
-  targets {
-    host_names: "localhost:6379"
-  }
-  [myprober.redis_probe] {
-    op: SET
-    key: "testkey"
-    value: "testval"
-  }
-}
-
 grpc_port: 9314
