+++ cloudprober/myprober2.cfg
@@ -1,0 +1,11 @@
+probe {
+  name: "redis_set"
+  type: HTTP
+  interval_msec: 10000
+  timeout_msec: 5000
+  targets {
+    host_names: "localhost:6379"
+  }
+}
+
+grpc_port: 9314
